Moved PreprocessSketchWithArduinoPreprocessor into proper place
